requirements.txt
fastapi
uvicorn
grpcio
grpcio-tools
boto3
python-dotenv
